modified mod function's return value to double.

+++ src/main/java/edu/handong/csee/pp1/bigdata/lab1_1/Calculator.java
@@ -112,8 +112,8 @@ public class Calculator {
 		return quot;
 	}
 
-	int mod(int first, int second) {
-		int mod = (int)Double.NaN;
+	double mod(int first, int second) {
+		double mod = Double.NaN;
 		try {
 			mod = first%second;
 		} catch (ArithmeticException e) {
